Move static constructor. PR comment: https://github.com/hyperledger/besu/pull/97#discussion_r333349309
