+++ besu/src/main/java/org/hyperledger/besu/services/BesuEventsImpl.java
@@ -17,8 +17,12 @@ package org.hyperledger.besu.services;
 import org.hyperledger.besu.ethereum.eth.sync.BlockBroadcaster;
 import org.hyperledger.besu.ethereum.eth.sync.state.SyncState;
 import org.hyperledger.besu.ethereum.eth.transactions.TransactionPool;
+import org.hyperledger.besu.plugin.data.BlockHeader;
 import org.hyperledger.besu.plugin.data.BlockPropagated;
+import org.hyperledger.besu.plugin.data.Quantity;
 import org.hyperledger.besu.plugin.services.BesuEvents;
+
+import java.util.function.Supplier;
 
 public class BesuEventsImpl implements BesuEvents {
   private final BlockBroadcaster blockBroadcaster;
@@ -38,8 +42,7 @@ public class BesuEventsImpl implements BesuEvents {
   public long addBlockPropagatedListener(final BlockPropagatedListener listener) {
     return blockBroadcaster.subscribePropagateNewBlocks(
         (block, totalDifficulty) ->
-            listener.onBlockPropagated(
-                BlockPropagated.of(block::getHeader, () -> totalDifficulty)));
+            listener.onBlockPropagated(blockPropagated(block::getHeader, () -> totalDifficulty)));
   }
 
   @Override
@@ -77,5 +80,21 @@ public class BesuEventsImpl implements BesuEvents {
   @Override
   public void removeSyncStatusListener(final long listenerIdentifier) {
     syncState.removeSyncStatusListener(listenerIdentifier);
+  }
+
+  private static BlockPropagated blockPropagated(
+      final Supplier<BlockHeader> blockHeaderSupplier,
+      final Supplier<Quantity> totalDifficultySupplier) {
+    return new BlockPropagated() {
+      @Override
+      public BlockHeader getBlockHeader() {
+        return blockHeaderSupplier.get();
+      }
+
+      @Override
+      public Quantity getTotalDifficulty() {
+        return totalDifficultySupplier.get();
+      }
+    };
   }
 }
